telem and failsafe working properly

+++ STM32Cube project files/Core/Inc/CRSF.h
@@ -10,7 +10,7 @@
 extern uint8_t rx_buffer[64*2];
 extern uint8_t tx_buffer[64];
 extern crsf_channels_t channel_data;
-extern volatile uint16_t  last_packet_received_time;
+extern volatile uint32_t  last_packet_received_time;
 extern volatile bool new_packet_recieved ;
 void crsf_init();
 uint8_t calculateCRC(uint8_t* buffer , int start_index , int bytes);
